refactor: migrate nativeKeydown to hotkey
