+++ packages/blocks/src/__internal__/utils/std.ts
@@ -206,3 +206,11 @@ export function maxBy<T>(items: T[], value: (item: T) => number): T | null {
 
   return maxItem;
 }
+
+export function isControlledKeyboardEvent(e: KeyboardEvent) {
+  return e.ctrlKey || e.metaKey || e.altKey;
+}
+
+export function isPrintableKeyEvent(event: KeyboardEvent): boolean {
+  return event.key.length === 1 && !isControlledKeyboardEvent(event);
+}
